ai-services/requirements.txt
# DataTrack AI Services - Python Dependencies
# OCR & Translation Services for KMRL Document Processing

# Core FastAPI and server dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Google Cloud APIs for OCR and Translation
google-cloud-vision>=3.4.0
google-cloud-translate>=3.11.0
google-auth>=2.17.0

# Data handling and validation
pydantic>=2.0.0
python-dateutil>=2.8.2

# File processing
Pillow>=10.0.0
python-magic>=0.4.27

# Development and debugging
python-dotenv>=1.0.0
